Refactor DefaultBinzImporter (extract the game specific importers into their own files)
